HFVHANDLER-295, clean up compiling warnings - not including the warnings in gctp and the parser.

+++ HDF5CFArray.cc
@@ -213,7 +213,7 @@ void HDF5CFArray::read_data_NOT_from_mem_cache(bool add_mem_cache,void*buf) {
 
         BESDEBUG("h5","Coming to use disk cache "<<endl);
 
-        long disk_cache_size = HDF5RequestHandler::get_disk_cache_size();
+        unsigned long long disk_cache_size = HDF5RequestHandler::get_disk_cache_size();
         string diskcache_dir = HDF5RequestHandler::get_disk_cache_dir();
         string diskcache_prefix = HDF5RequestHandler::get_disk_cachefile_prefix();
 
@@ -1311,7 +1311,7 @@ void
 HDF5CFArray::write_data_to_cache(hid_t dset_id, hid_t /*dspace_id*/, hid_t /*mspace_id*/, hid_t memtype,
     const string& cache_fpath, short dtype_size, const vector<char> &buf, int nelms) {
 
-    long disk_cache_size = HDF5RequestHandler::get_disk_cache_size();
+    unsigned long long disk_cache_size = HDF5RequestHandler::get_disk_cache_size();
     string disk_cache_dir = HDF5RequestHandler::get_disk_cache_dir();
     string disk_cache_prefix = HDF5RequestHandler::get_disk_cachefile_prefix();
     HDF5DiskCache *disk_cache = HDF5DiskCache::get_instance(disk_cache_size,disk_cache_dir,disk_cache_prefix);
