+++ HDF5RequestHandler.h
@@ -116,11 +116,11 @@ class HDF5RequestHandler:public BESRequestHandler {
     static bool get_use_disk_cache() {return _use_disk_cache;}
     static string get_disk_cache_dir() { return _disk_cache_dir;}
     static string get_disk_cachefile_prefix() { return _disk_cachefile_prefix;}
-    static long get_disk_cache_size() {return _disk_cache_size;}
+    static unsigned long long get_disk_cache_size() {return _disk_cache_size;}
     static bool get_disk_cache_comp_data() { return _disk_cache_comp_data;}
     static bool get_disk_cache_float_only_comp(){return _disk_cache_float_only_comp_data;}
     static float get_disk_comp_threshold() {return _disk_cache_comp_threshold;}
-    static long get_disk_var_size() {return _disk_cache_var_size; }
+    static unsigned long get_disk_var_size() {return _disk_cache_var_size; }
 
     static bool get_use_eosgeo_cachefile() {return _use_latlon_disk_cache; }
     static string get_latlon_disk_cache_dir() { return _latlon_disk_cache_dir;}
@@ -165,11 +165,11 @@ class HDF5RequestHandler:public BESRequestHandler {
      static bool _use_disk_cache;
      static string _disk_cache_dir;
      static string _disk_cachefile_prefix;
-     static long _disk_cache_size;
+     static unsigned long long _disk_cache_size;
      static bool _disk_cache_comp_data;
      static bool _disk_cache_float_only_comp_data;
      static float _disk_cache_comp_threshold;
-     static long _disk_cache_var_size;
+     static unsigned long _disk_cache_var_size;
 
          
      static bool _use_disk_meta_cache;
